requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
openai>=1.0.0

# Web framework
flask>=3.0.0
flask-cors>=4.0.0

# Data handling
pydantic>=2.0.0
python-dotenv>=1.0.0

# Database and logging
sqlalchemy>=2.0.0
alembic>=1.13.0

# Visualization and analysis
plotly>=5.18.0
pandas>=2.0.0

# Utilities
python-dateutil>=2.8.0

